Update JCloudsStorageModule.java
getSizeInBlocks() returns the virtual blocks in the cloud side. There are 1048576 virtual blocks as stated in JCloudsStorageModule construtor, and there are 1048576/ BLOCK_IN_CLUSTER clusters.
So size in virtual blocks should be  mNumberOfCluster * BLOCK_IN_CLUSTER.Cluster and blocks are confusing,as i am using s3 api.cluster is equal to object in s3.